pyproject.toml
[project]
name = "VoiceTranscription-telegram-bot"
version = "1.0.0"
description = """A simple Telegram bot built with Python that automatically transcribes
voice messages and audio files into text, then generates concise summaries"""
authors = [
    {name = "Andrii Sonsiadlo", email = "andrii.sonsiadlo@gmail.com"},
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "groq>=0.34.0",
    "python-dotenv>=1.2.1",
    "python-telegram-bot>=22.5",
]

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "ruff>=0.14.5",
]
